Join literal sequence Expr logic into Exprs
Currently we have `Expr.ofSeq` and `ExprSeq.unapply` to handle such cases.
Instead we provide a single `Exprs` that contains an `apply` and `unapply`.

Furthermore having both `Expr.ofSeq` and `Expr.ofList` has lead to confusions on which one to use for varargs.
This will help guide users to the correct logic to use for varargs.