Store annotation arguments as untyped tree
We can't properly type annotation arguments which are empty array
literals because we would need the corresponding parameter type which
would require forcing the annotation constructor early. The previous
commit worked around this by using `WildcardType` as the element type
but that's a hack. Instead, we can fix the problem by always storing
annotation arguments as untyped trees and using `applyOverloaded` to
type them when the annotation tree is forced. This requires some
refactoring to make `applyOverloaded` work both with typed and untyped
arguments (the variants are now accessible using `untpd.applyOverloaded`
and `tpd.applyOverloaded`).

(the abstract `def FunProto` in Trees that gets implemented by
forwarding in untpd/tpd sounds like a good candidate for `export`,
except we can't export a class constructor)