Switch to implementation scheme described in PR
The previous scheme was more complicated and looks more fragile,
in particular if complex staging operations are applied to a retained
inline method. In that case we have to rely on the fact that the rhs
of a retained inline method is in fact its inlineable body, and not
the retained body.