Rewrite classOf handling
So far, a miniphase took care of rewriting Predef.classOf to a class
constant, but this means that `classOf` could survive in annotations
(since they're not transformed) and needed to be handled by the backend.
To simplify things, we take advantage of the constant folding done in
the typer to directly replace `classOf` calls by a constant.