Do not rely on `ArrayIndexOOBE` in `Vector.{head,last}`.
This is necessary for Scala.js, where
`ArrayIndexOutOfBoundsException`s are Undefined Behavior, and
cannot be reliably caught.

Those methods throw a `NoSuchElementException` for an empty vector,
and those exceptions have never been considered UB in Scala.js
before. It is therefore a breach of contract for
`Vector.{head,last}` to become UB on empty vectors.

There are many other places where we catch AIOOBEs, but in those
other places we throw an `IndexOutOfBoundsException` instead.
While it is not quite correct to turn them into UBs either, at
*least* there is precedent for OOBEs to be considered UB in
Scala.js. So this commit does not change them, being afraid of the
consequences to JVM performance.

It *is* still bad, though (index out of bounds on Scala collections
were never UB before), but at least it is a bit less bad than
before.