Better cleanup strategy for GH Actions CI
Before each GH Actions job, remove everything from:

- The working directory (where the Dotty repo is cloned)
- The temp directory
- The user home directory

These directories are mounted by Docker before the workflow
is started and hence can be persisted between different CI
runs and we don't want that. The list of the directories can be obtained by looking at the `docker create` command
that the "Initialize containers" step runs for each job of
the CI. Only volumes that are vital for CI functioning are
spared during the cleanup.