Fix #8619: Demand `transparent` for whitebox inlines
Previously, whitebox inlines could be declared in three different ways

 1. writing `inline def f <: T = ...` for inline methods
 2. writing `inline given f: _ <: T` = ...` for inline givens
 3. leaving out the result type for inline methods

The last was an oversight, not backed by the spec. It turned out that it
was not that easy to implement (3), so it was not done, and afterwards
code exploited the loophole.

In the new scheme, `whitebox` inlines demand `transparent`. The result type
can be given or left out, the effect is the same. The old `<:` result type
syntax will be phased out in a subseqent PR once the new syntax is in a release.

`transparent` is a soft modifier. It is valid only together with `inline`. Why not
allow `transparent` on its own and let it subsume `inline`. The point is that
we should steer users firmly towards blackbox macros, since they are much
easier to deal with for tooling. This means we want to make whitebox macros
strictly more verbose than blackbox macros. Otherwise someone might find
`transparent` "nicer" than `inline` and simply use it on these grounds without
realizing (or caring about) the consequences.

For the same reason I did not follow the (otherwise tempting) idea to simply
re-use `opaque` instead of `transparent`. An opaque inline kleeps its type on
expansion whereas a transparent one gets the type of its expansion. But that
would have nudged to user to prefer `inline` over `opaque inline`, so would
again have gjven the wrong incentive.

On the other hand, `transparent` as a dual of `opaque` is nice. It fits into
the same terminology. It's simply that type aliases are transparent by default
and have to be made opaque with a modifier, whereas inline methods are opaque
by default, and have to be made transparent by a modifier.