Don't write `(using ctx: Context)`
The new way we organize context passing has one downside. We start with `(using Context)`
because it is less cluttered, but then if in the body we have to refer to the current
context we go back to `(using ctx: Context)`. I havealso seen occurrences of `(using ctx: Context)`
without an actual usage of `ctx` in the body.

This going back and forth is awkward. Fortnately, there's an alternative: `Contexts` contains
a method `curCtx` which summons the current context. We should always use `curCtx` instead of
`ctx`, which means we never need to write `(using ctx: Context)`.

Or maybe we can even go back renaming `curCtx` to `ctx`? Then we could simply drop all `ctx` names
in parameters and hopefully get the same meaning and performance.