Use indentation to control newline insertion before parentheses
Scala currently uses an ad-hoc set of rules to determine whether a line
starting with an opening `(`, `[`, or `{` is an argument to the previous
line or starts a new statement. Specifically, a new line is always
inserted but is then skipped if the next token is a `{`. On the other
hand, lines starting with `[` and `(` always start a new statement.
This leads to the gotchas that
```
val x= foo
{ ... }
```
treats the braces of an argument to `foo`. A blank line is required to avoid this.
On the other hand,
```
foo(arg1)
  (arg2)
  (arg3)
```
does not work; the following lines are interpreted as separate statements.

In this PR, indentation is used instead to determine whether a following line starting
with `(`, `[`, ir `{` is an argument or a separate statement. In the first example, it
is a separate statement, since it is not indented. In the second example, it is treated
as an argument, since it is indented.

If we are past a blank line, we always assume the newline. For instance, this
example from Pouring.scala works:
```scala
  val moves =
    val glasses = 0 until capacity.length

	   (for g <- glasses yield Move.Empty(g))
    ++ (for g <- glasses yield Move.Fill(g))
    ++ (for g1 <- glasses; g2 <- glasses if g1 != g2 yield Move.Pour(g1, g2))
```
Without the blank line, the `(for g <- ...` expression would be treated as an additional
argument to `capacity.length`.