Remove incorrect type simplification
`TypeOps#simplify` is supposed to produce a type `=:=` the original
type, but it replaced type variables created using `newTypeVar` by one
of their bound, this is wrong since that type variable might get
instantiated to a different type later. This simplification was added
in #2209 to prevent a pickling crash in tests/pos/i2152.scala but does
not appear to be needed anymore: an uninstantiated type variable being
preserved after typer is always wrong, so the underlying bug must have
been fixed since then.