Handle dependent parameters hidden behind a TypeVar
This is needed for type inference with a type variable selection (added
in the next commit of this PR) to work properly, but it also affects master.

Because dependencies can now span multiple TypeLambdas,
`addToConstraint` had to be adjusted to properly propagate all
constraints.

i5735 doesn't compile anymore but I don't think it ever really made
sense.