Do singleton widening when migrating to Scala 3.
The previous changes failed the community build for squants. What happened was
that an implicit in the package object of squants was not found in motion/AngularAcceleration.

Normally, the context object would be found in the normal implicit context of the use site.
But AngularAcceleration was defined as
```scala
package squants.motion
```
which meant that `squants` by itself was not in the implicit scope. If it had been defined like this:
```scala
package squants
package motion
```
it would have worked fine. So the implicit hit it relied on the fact that under -source:3.0-migration
we also search package prefixes of types. Only, I believe there were no such prefixes normally since the types
in question started with `this[motion]` so no mention of `squants`. But if we widen the term refs on
the paths then we mention something that mentions in turn `scala` and things are OK. I did not track
down what this was, but that's what must have happened.

So, to keep compatibility with old code, we should still enable this under -source:3.0-migration.