Avoid time travel when handling of LazyRefs
The idea of LazyRefs is that they are always completed in the current context.
But that did not work if the LazyRef is mapped in a type map. Here, the TypeMap's
context was used to do the remap. We now use the context that forced the LazyVal
instead.

The problem manifested itself with failing replTests when the new implicitScope
scheme was used.