Also include companions of prefixes in implicit scope
In that way, we include all of the implicit scope of Scala 2, except
for package objects. The idea is that everything that is reasonable
should be included in the implicit scope.

Package objects were excluded since they had a dual purpose:

 1. as a way to define implicits that are in scope for a whole project
 2. as a prefix for implicits of all types coming out of that project

It's often that one wants only (1) but then is surprised to also get (2).