Fix #9132: Align with Scala 2's handling of () infix arguments
A right argument of an infix operation is kept as an empty tuuple, instead
of being desugared to an empty parameter list.

The most involved part of the change was getting decent error messages, since
it required propagating the information that an apply is an infix all the way
into application checking. Hopefully, we can profit from the change when
we drop auto-tupling, since then similar info will be needed for error
messages in more situations.