Fix context for unpickling symbol annotations
For member definitions, the owner for unpickling
annotations should be the owner of the symbol.

Note that the type checking has a more complex mechanism
to avoid entering argument definitions in the enclosing class.
See `Typer.annotContext` and `Context.exprContext`.

In the context of unpickling, it is unclear how to get
that mechanism to work.