Improve healing failure message
When healing is attempted and quote.Type is not available, the levels are irrelevant for the user.
The users need to know that this quote.Type is missing and can fix the issue by adding it. Otherwise, they need to
know that a type that was not supposed to be in the quotes is used within them.