Unified extension methods
Dotty currently uses three syntactically different schemes for extension methods

 1. Single methods with prefix parameters such as `def (x: T).f(y: U)` or `def (x: T) op (y: U)`.
    These methods can be abstract, can override each other, etc, which is crucial for infix ops
    in type classes.
 2. Given instances defining (only) extension methods with syntax `extension ops { ... }`
 3. Collective extension methods with syntax `extension on (x: T) { ... }`

This PR proposes another scheme that unifies (1) and (3) and all but eliminates the need for (2).

The simplification is achieved by redefining many fundamentals of extension methods. In a sense,
the price we pay for user-facing simplifications is a more complex implementation scheme. But that's
probably a price worth paying.

The first commit is docs only. I updated the doc pages for extension methods, type classes, and
the grammar.