Fix ApproximateGadtAccumulator
Explanations are in the relevant doc.

Provisionally restored GADT-approximating version of `recover` in
`adaptToSubType`, as it turns out to be necessary to typecheck i7044.scala.

Moved boundspropagation.scala to pending.
Originally, ApproximateGadtAccumulator allowed it to typecheck, so it
was added as a test. Since `recover` was provisionally restored, the
test succeeds again. However, as GADT approximation is only meant to fix
GADT member selection issues and the example does not involve those,
we probably don't want it as a pos test. At the same time, it should
typecheck, so we shouldn't delete it either.