Scala.js: Support for reflective calls.
This is significantly different from the code in scalac. In scalac,
reflective calls reach the back-end as dedicated `ApplyDynamic`
nodes. In dotc however, they are expanded by the type-checker into
user-level calls to method of `scala.reflect.Selectable`.

We must therefore reverse-engineer the patterns of code produced by
the compiler to construct SJSIR `Apply` nodes with reflective proxy
method names.

The big comment on `JSCodeGen.genReflectiveCall` explains the
translation in detail.