Better handling of ctrl-c while the compiler is running
Previously, the compiler tried its best to continue working after
ctrl-c, which usually resulted in a flood of error messages. This commit
takes inspiration from https://github.com/scala/scala/pull/6479 to
handle ClosedByInterruptException and InterruptedException gracefully
and avoid this.

Note that the Scala 2 PR goes further: it checks `Thread.interrupted()`
to set a global `cancelled` flag (Should we do the same with our
`Run#isCancelled` ?) and it also catches InterruptedException at the
top-level so that ctrl-c does not end up printing a stacktrace, but this
is beyond what I have the time to look at currently.