Scala.js: Fix `classOf[Null]` and `classOf[Nothing]`.
For almost all purposes, `Null` and `Nothing` should be erased to
`scala.runtime.{Null$,Nothing$}`, so we directly change
`encodeClassName` to do that.

The only place where they should be encoded as the special IR
TypeRefs `NullRef` and `NothingRef` is in param and result types of
`MethodName`s, which was already taken care of by
`paramOrResultTypeRef` but that was adjusted to compensate for the
changes in `encodeClassName`. The new `paramOrResultTypeRef` is now
equivalent to the one in scalac.