Translate Java varargs `...T` into `T*` instead of `(T & Object)*`
This is much more convenient for users but is more complicated for the
compiler since we might still need to translate the varargs into an
`Object[]` in bytecode. Since ElimRepeated was already responsible for
doing some adaptation of repeated arguments, it now also takes care of
this (unlike Scala 2 where this is handled in Erasure).

Fixes #9439.