Decouple quoted Type interface from encoding
Before, `quoted.Type` was defined as

```scala
// old
class Type[X <: AnyKind]:
  type T = X
```

Where `T` is the proper type of the instance of a `Type`. This is the type used for all operations within the compiler.

The only purpose of the type `X` was to be able to write context bounds as `[X: Type]` and staged types `Type[X]` instead of
`Type { type T = X }`.
Unfortunately, the type `X` makes the assumption that `X` is known, which implies that instances of `Type` with unknown types need to be encoded with a whildcard type.
A `val t: Type[?] = ...` is unusable as we cannot heal staged wildcard types, even though we do have a `t.T` which can be handled.

Now, we change the definition of `Type` to only have the type member.
```scala
// new
class Type:
  type T <: AnyKind
```

This solves the aforementioned limitations with unkown staged types.
Now we are also able to remove the wildcard from `QuoteContext.tasty.Type.seal` and the result can be used in quotes.

To keep the context bound syntax and provide more descriptive name for the concept, we defined the `quoted.Staged` type alias.

```scala
type Staged[X <: AnyKind] = Type { type T = X }
```