Always override enumLabel in each enum case.
As a rule, even class cases override it to be consistent with
singleton cases which must have a value matching the identifier.

this also moves runtime.EnumValues into src-bootstrapped so that
fromName can directly call .enumLabel

This commit also removes the forced override of productPrefix in
EnumValue and individial cases, this allows the user
to customise productPrefix as with typical case classes.
An override of productPrefix will be generated in SyntheticMembers
to forward to enumLabel if a class parent does not define one.