Scala.js: Add neg tests for bad uses of `reflect.Selectable`.
We add some infrastructure to be able to perform neg tests that are
specific to Scala.js. Such tests require the scalajs-library jar
and the dotty-libraryJS jar on the classpath, and must be compiled
with the `-scalajs` option.

We use this infrastructure to test existing compile errors reported
for unsupported uses of `reflect.Selectable` in Scala.js.