Remove Reflection.TypeOrBounds abstraction
* Align with TASTy format
* Make it simpler to handle types with the reflection API
* Remove ParamInfo from LambdaType (it was unused)

This abstraction is there for historical reasons only, from the time when there was split between TypeTree, TypeBoundsTree and Tree. This distinction is no longer in the API and the types should follow the same design.

Migration help: If there match on `Type` that matched agaist a `TypeOrBound`
```scala
def f(t: TypeOrBound) =
  t match
    case t: Type =>
    ...
```
The cases for `TypeBounds` and `NoPrefix` should be added before the `Type` test

```scala
def f(t: Type) =
  t match
    case t: TypeBounds =>
    case NoPrefix() =>
    case t => // t: Type
    ...
```