Fix #8769: change dottydoc root to docs/ subdirectory
Previously it defaulted to the project root which lead to various
issues, notably it meant that the whope project ended up being copied in
_site when running `sbt doc`, what's even more fun is that when I tried
using sbt 1.4.0-RC1 this caused a compiler crash because it ended up
trying to copy a socket file that sbt had created in `target/`.

Usincs "docs/" like this still doesn't really make sense when multiple
subprojects exist since we'll try using the same directory for all of
them, but it's less worse than what we had before and it doesn't seem
worth investing more time in dottydoc since it's getting replaced.