Fix #9775: Fix primitive-in-condition handling in backend
isPrimitive/getPrimitive are overloaded: they both have a version which
takes a symbol as input and a version which takes a tree, the latter can
handle array get and set primitives which map to different JVM
instructions depending on their type, but we don't care about that when
emitting conditions in genCond: we only need to look for primitives
which correspond to boolean operations and those are already handled by
the symbol versions of isPrimitive/getPrimitive. By using these methods
in genCond we avoid running some unnecessary code and we fix #9775 which
was triggered because the `getPrimitive` overload we were using required
forcing `lhs`. This change also aligns us with what the scalac backend
does.