Make only Wiki-style links supported in Md syntax
Also add slash-escaping to Wiki links.

There's unfortunate syntax conflicts between the link-to-definition syntax and
Markdown link syntax. In particular:

+ The former uses parentheses for overload resolution. These would always need
  to be backslash-escaped, which raises questions on how to differentiate
  between parentheses escaped from Markdown syntax and parentheses escaped from
  link syntax. AFAICT it's not really possible to distinguish between the two.
+ In Markdown one can write [header](#header) to link to a local header. This
  conflicts with definition link syntax for linking to a local member. Already
  existing Markdown documents can potentially be using that link syntax, and
  breaking those links would be unfortunate.

Also, writing [](link.to.def) in Markdown is just plain ugly. None of those
problems are impossible to live with, but I think things will just be nicer if
we require using Wiki link syntax to link to definitions.