Fix @switch warnings for matches on value classes
Matches on value classes that have an underlying switchable type may be
emitted as tableswitch or lookupswitch, but the shape of the result tree
differs from ordinary switch-compiled matches.

Previously, this caused a spurious warning to be issued if such a match
was annotated with @switch, as none of the result cases were discovered
by the warning logic, and hence the match was warned as having too few cases.

With the warning for too few cases now removed, we have the opposite
issue: in no circumstance is a switch warning issued for @switch
annotated matches on value classes.

This commit attempts to address this issue and restore the @switch
warnings for those matches on value classes where a tableswitch or
lookupswitch is not emitted.

A complicating factor is that the original case types for matches
on value class extractors are not singleton types, and so counting
the number of unique types is not useful for determining the number
of original cases.