Remove quoted.Lambda extractor
This extractor is by nature unsound because it ignores argument semantics.
`Expr.betaReduce` provides a sound version of this conversion with a slightly diferent API.

The error messages for this extractors where never good due to the complexity of its signature.

This kind of unsafe transformations should be performed on the reflection API instead.