An alternative scheme for precise apply methods of enums
Keeps many elements from #9922 but the modality where we do the
widening is different. The new rule is as follows:

In an application of a compiler-generated apply or copy method of an enum case,
widen its type to the underlying supertype of the enum case by means of a type
ascription, unless the expected type is an enum case itself.