Warn users on non canonical type references
Help users write the correct variant of the quoted type encoding.
Explicit construction of a `scala.quoted.Type[_]` should be done with its constructor.
Explicit construction of types in not encouraged as it should be created implicitly.
Creating an `Type` explicitly is only useful for debugging purposes.
Likewise, we encourage the use of the explicit inner type reference.