Remove Inlined trees before erasure
In this first step we remove the Inlined trees in FirstTransform just after unpickling.
We also remove the YCheckPositions wich checked that all changes in sources where wrapped in Inline nodes.
This invariant does not hold after FirstTransform anymore.
YCheckPositions was added to make sure that the transformation fases kept the Inlined nodes until erasure which is not needed anymore.