WIP Use InPlaceCopier
This does not work yet. The error is in lambda lift, where we traverse the same
tree several times in different methods. The shared trees are snippets generated
by lazyvals. The snippets need to be substituted, mapping original references
to proxy references. But the proxy references are different in different methods.
The proxy reference itself is generated by lambdaLift's copier which is a regular
copier. But the enclosing applicatiin is copied by the in-place copier, and that
way we get mixed up.

I don't really see a way to fix this.